% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{draw.modSim}
\alias{draw.modSim}
\title{Draw Heatmap of modSim object}
\usage{
draw.modSim(x, k, ind = NULL, ...)
}
\arguments{
\item{x}{modSim object}

\item{k}{number of split for heatmap's columns}

\item{ind}{number of columns to keep}

\item{...}{supplementary parameters}
}
\value{
draw x
}
\description{
Draw Heatmap of modSim object
}
\examples{
library(survMS)
res_paramW = get_param_weib(med = 2.5, mu = 1.2)
listCoxSimCor_n500_p1000 <- modelSim(model = "cox", matDistr = "mvnorm", 
                                     matParam = c(0,0.6), n = 500, 
                                     p = 1000, pnonull = 20, betaDistr = 1, 
                                     hazDistr = "weibull", 
                                     hazParams = c(res_paramW$a, res_paramW$lambda), 
                                     seed = 1, d = 0)
print(listCoxSimCor_n500_p1000)
hist(listCoxSimCor_n500_p1000)
#draw(listCoxSimCor_n500_p1000, k = 3)
}
