% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_exp}
\alias{get_param_exp}
\title{Getting parameters of exponential distribution}
\usage{
get_param_exp(int_a = c(1e-06, 110), med, mu)
}
\arguments{
\item{int_a}{: grid of first parameter of the distribution}

\item{med}{: median of real dataset}

\item{mu}{: mean of real dataset}
}
\value{
: list with parameter values of the distribution
}
\description{
Getting parameters of exponential distribution
}
\examples{
library(survMS)
}
