% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_ln}
\alias{get_param_ln}
\title{Getting parameters of log-normal distribution}
\usage{
get_param_ln(var = 170000, mu = 2325)
}
\arguments{
\item{var}{: variance of real dataset}

\item{mu}{: mean of real dataset}
}
\value{
list with parameter values of the distribution
}
\description{
Getting parameters of log-normal distribution
}
\examples{
library(survMS)
}
