% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param_ln2}
\alias{get_param_ln2}
\title{Getting parameters of log-normal distribution}
\usage{
get_param_ln2(int_a = c(0.1, 11), med, mu)
}
\arguments{
\item{int_a}{: grid of first parameter of the distribution}

\item{med}{: median of real dataset}

\item{mu}{: mean of real dataset}
}
\value{
: list with parameter values of the distribution
}
\description{
Getting parameters of log-normal distribution
}
\examples{
library(survMS)
}
