% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSim.R
\name{Heatmap.modSim}
\alias{Heatmap.modSim}
\title{Heatmap of Covariate matrix}
\usage{
\method{Heatmap}{modSim}(x, k, ind = NULL, ...)
}
\arguments{
\item{x}{output of modelSim function (must be of type modSim)}

\item{k}{number of column split}

\item{ind}{indices of columns to keep}

\item{...}{supplementary parameters}
}
\value{
Heatmap x
}
\description{
Heatmap of Covariate matrix
}
\examples{
library(survMS)
res_paramW = get_param_weib(med = 2.5, mu = 1.2)
listCoxSimCor_n500_p1000 <- modelSim(model = "cox", matDistr = "mvnorm", 
                                     matParam = c(0,0.6), n = 500, 
                                     p = 1000, pnonull = 20, betaDistr = 1, 
                                     hazDistr = "weibull", 
                                     hazParams = c(res_paramW$a, res_paramW$lambda), 
                                     seed = 1, d = 0)
print(listCoxSimCor_n500_p1000)
hist(listCoxSimCor_n500_p1000)
Heatmap.modSim(listCoxSimCor_n500_p1000, k = 4)
}
