% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{Heatmap}
\alias{Heatmap}
\title{Internal heatmap object}
\usage{
Heatmap(x, k, ind = NULL, ...)
}
\arguments{
\item{x}{modSim obkect}

\item{k}{number of clusters for the heatmap}

\item{ind}{number of column covariate matrix}

\item{...}{supplementary arguments}
}
\value{
Heatmap x
}
\description{
Internal heatmap object
}
\examples{
library(survMS)
}
