% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSim.R
\name{hist.modSim}
\alias{hist.modSim}
\title{Histogram of survival times}
\usage{
\method{hist}{modSim}(x, ...)
}
\arguments{
\item{x}{output of modelSim function (must be of type modSim)}

\item{...}{supplementary parameters}
}
\value{
hist x
}
\description{
Histogram of survival times
}
\examples{
library(survMS)
}
