% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survMS.R
\docType{package}
\name{survMS}
\alias{survMS}
\title{survMS Package}
\description{
Package enables simulating data with different levels of complexity from  survival models (Cox, AFT and AH models).
}
\author{
Mathilde Sautreuil \email{mathilde.sautreuil@gmail.com}
}
