% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{SurvTimesAFTshiftLN}
\alias{SurvTimesAFTshiftLN}
\title{Simulation survival times from Shift AFT/Log-normal model}
\usage{
SurvTimesAFTshiftLN(Z, beta, beta2, Y, pp, hazParams)
}
\arguments{
\item{Z}{Matrix of covariates}

\item{beta}{regression parameter}

\item{beta2}{vector of regression parameter or distribution of regression parameter}

\item{Y}{random uniform}

\item{pp}{number of pertinent covariates}

\item{hazParams}{distribution parameters of baseline hazard risk}
}
\value{
Ts Observed times
}
\description{
Simulation survival times from Shift AFT/Log-normal model
}
\examples{
library(survMS)
}
\keyword{internal}
