% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{SurvTimesCoxLN}
\alias{SurvTimesCoxLN}
\title{Simulation survival times from Cox/Log-normal model}
\usage{
SurvTimesCoxLN(Z, beta, Y, pp, hazParams)
}
\arguments{
\item{Z}{Matrix of covariates}

\item{beta}{regression parameter}

\item{Y}{random uniform}

\item{pp}{number of pertinent covariates}

\item{hazParams}{distribution parameters of baseline hazard risk}
}
\value{
Ts Observed times
}
\description{
Simulation survival times from Cox/Log-normal model
}
\examples{
library(survMS)
}
\keyword{internal}
