% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSim.R
\name{plot.modSim}
\alias{plot.modSim}
\title{Survival or hazard curves of simulated data}
\usage{
\method{plot}{modSim}(x, ind, type = "surv", ...)
}
\arguments{
\item{x}{output of modelSim function (must be of type modSim)}

\item{ind}{vector (individuals to show)}

\item{type}{type of plots (survival or hazard curves)}

\item{...}{supplementary parameters}
}
\value{
plot x
}
\description{
Survival or hazard curves of simulated data
}
\examples{
library(survMS)
ind = sample(1:500, 5)
}
