% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSim.R
\name{print.modSim}
\alias{print.modSim}
\title{Print information about data simulation}
\usage{
\method{print}{modSim}(x, ...)
}
\arguments{
\item{x}{output of modelSim function (must be of type modSim)}

\item{...}{supplementary parameters}
}
\value{
print x
}
\description{
Print information about data simulation
}
\examples{
library(survMS)

}
